requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
greenlet==3.2.1
idna==3.10
numpy==2.2.5
pandas==2.2.3
psycopg2==2.9.10
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wheel==0.45.1
